TDD iteration 6 - REFACTOR Add ReportingWithAccountsTest class
