+++ src/test/java/com/acme/banking/dbo/service/reporting/ReportingTest.java
@@ -34,85 +34,5 @@ public class ReportingTest {
                 .build();
     }
 
-    @Test
-    public void shouldGetReportWithNewLinesWhenBranchHasAccount() throws ClassNotFoundException {
-        //region when
-        when(stubBranch.getAccounts()).thenReturn(stubAccounts);
-        report = sut.getReport(stubBranch);
-        //endregion
 
-        //region then
-        assertThat(report).isEqualTo("# \n## " + stubClientId);
-        //endregion
-    }
-
-    @Test
-    public void shouldGetReportWithAccountIdWhenBranchHasAccount() throws ClassNotFoundException {
-        //region when
-        when(stubBranch.getAccounts()).thenReturn(stubAccounts);
-        when(stubAccount.getClientId()).thenReturn(stubClientId);
-        report = sut.getReport(stubBranch);
-        //endregion
-
-        //region then
-        assertThat(report).isEqualTo("# \n## " + stubClientId);
-        //endregion
-    }
-
-
-    @Test
-    public void shouldGetReportWhenBranchHasTwoAccounts() throws ClassNotFoundException {
-        //region given
-        UUID stubClientId2 = UUID.randomUUID();
-        stubAccounts = new MockitoAccountCollectionBuilder()
-                .withAccount(new MockitoClientBuilder().withId(stubClientId).build())
-                .withAccount(new MockitoClientBuilder().withId(stubClientId2).build())
-                .build();
-        //endregion
-
-        //region when
-        when(stubBranch.getAccounts()).thenReturn(stubAccounts);
-        report = sut.getReport(stubBranch);
-        //endregion
-
-        //region then
-        assertThat(report).isEqualTo("# \n## " + stubClientId + "## " + stubClientId2);
-        //endregion
-    }
-
-    @Test
-    public void shouldGetReportWhenBranchHasNoName() {
-        //region when
-        when(stubBranch.getName()).thenReturn(null);
-        report = sut.getReport(stubBranch);
-        //endregion
-
-        //region then
-        assertThat(report).containsOnlyOnce("# ");
-        //endregion
-    }
-
-    @Test
-    public void shouldGetReportWhenBranchIsEmpty() {
-        //region when
-        when(stubBranch.getName()).thenReturn("BrunchName");
-        report = sut.getReport(stubBranch);
-        //endregion
-
-        //region then
-        assertThat(report).containsOnlyOnce("# BrunchName");
-        //endregion
-    }
-
-    @Test
-    public void shouldGetReportWhenBranchClientsCollectionNotEmpty() throws ClassNotFoundException {
-        //region when
-        when(stubBranch.getAccounts()).thenReturn(stubAccounts);
-        report = sut.getReport(stubBranch);
-        //endregion
-
-        //region then
-        assertThat(report).isEqualTo("# \n## " + stubClientId);
-        //endregion
-    }
 }
